feat: move validation into its own package and reference in go.mod
