+++ cmd/flipt/validate.go
@@ -7,7 +7,7 @@ import (
 
 	"github.com/spf13/cobra"
 	"go.flipt.io/flipt/internal/containers"
-	"go.flipt.io/flipt/internal/cue"
+	"go.flipt.io/flipt/validation"
 	"go.flipt.io/flipt/internal/storage/fs"
 )
 
@@ -64,7 +64,7 @@ func (v *validateCommand) run(cmd *cobra.Command, args []string) error {
 		}
 
 		opts = append(opts, fs.WithValidatorOption(
-			cue.WithSchemaExtension(schema),
+			validation.WithSchemaExtension(schema),
 		))
 	}
 
@@ -74,7 +74,7 @@ func (v *validateCommand) run(cmd *cobra.Command, args []string) error {
 		_, err = fs.SnapshotFromPaths(logger, os.DirFS("."), args, opts...)
 	}
 
-	errs, ok := cue.Unwrap(err)
+	errs, ok := validation.Unwrap(err)
 	if !ok {
 		return err
 	}
