+++ validation/extended.cue
@@ -1,3 +1,4 @@
 #Flag: {
 	description: =~"^.+$"
 }
+
