+++ validation/testdata/invalid_yaml_stream.yaml
