+++ validation/testdata/valid_yaml_stream.yaml
