+++ validation/validate_fuzz_test.go
@@ -1,7 +1,7 @@
 //go:build go1.18
 // +build go1.18
 
-package cue
+package validation
 
 import (
 	"bytes"
